CHANGE - TPlaceholdersManipulation trait created (placeholders replacement and manipulation) - TPlaceholdersManipulation implemented in PdfCreator
